fix: create subgraph when multiple roots present
If there is more then one root present in the config
we need to create a subgraph that only includes the
Descendants of the desired root and omits the rest.
If we try to step through the initial dag instead it will
create a locking situation that will break execution.

Signed-off-by: Marvin Drees <marvin.drees@9elements.com>